requirements.txt
aiochclient==1.2.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.11
django-cacheops==4.1
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-jsonfield==1.3.1
django-mptt==0.10.0
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.8
djangorestframework==3.9.2
drf-nested-routers==0.91
drf-writable-nested==0.5.1
funcy==1.13
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.6.1
numpy==1.17.4
pandas==0.25.3
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
yarl==1.4.1
django-concurrency==2.1.1
